Rename distance in Similarity.cs, simplify and rename image repository and change type of distance search parameter to decimal

+++ Api/Implementations/Commons/SearchService.cs
@@ -26,7 +26,7 @@ public class SearchService : ISearchService
     }
 
     public async Task<IEnumerable<IGrouping<byte[], File>>> SearchAsync(string[] hypotheticalDuplicates,
-        FileSearchType searchType, PerceptualHashAlgorithm? perceptualHashAlgorithm, int degreeOfSimilarity = 0,
+        FileSearchType searchType, PerceptualHashAlgorithm? perceptualHashAlgorithm, decimal degreeOfSimilarity = 0,
         CancellationToken cancellationToken = default)
     {
         switch (searchType)
